+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers.NUnit.Integration/ArtifactHelpers/ClientHelperIntegrationTests.cs
@@ -1,26 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using kCura.Relativity.Client;
 using NUnit.Framework;
 using Relativity.API;
 using Relativity.Services.ServiceProxy;
 using Relativity.Test.Helpers.ArtifactHelpers;
-using Relativity.Test.Helpers.ServiceFactory.Extentions;
 using Relativity.Test.Helpers.SharedTestHelpers;
-using Relativity.Test.Helpers.WorkspaceHelpers;
 
 namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 {
 	[TestFixture]
 	public class ClientHelperIntegrationTests
 	{
-		private IHelper testHelper;
+		private IHelper _testHelper;
 		private IServicesMgr _servicesManager;
 		private Services.ServiceProxy.ServiceFactory _serviceFactory;
-		private IRSAPIClient _client;
 		private int _clientArtifactId;
 		const string _clientName = "TestClientName";
 
@@ -33,8 +26,7 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 			{
 				configDictionary.Add(testParameterName, TestContext.Parameters[testParameterName]);
 			}
-			testHelper = new TestHelper(configDictionary);
-			_client = testHelper.GetServicesManager().GetProxy<IRSAPIClient>(ConfigurationHelper.ADMIN_USERNAME, ConfigurationHelper.DEFAULT_PASSWORD);
+			_testHelper = new TestHelper(configDictionary);
 			_serviceFactory = GetServiceFactory();
 		}
 
@@ -43,19 +35,18 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 		{
 			if (_clientArtifactId != 0)
 			{
-				Client.Delete_Client(_serviceFactory, _clientArtifactId);
+				ClientHelper.DeleteClient(_serviceFactory, _clientArtifactId);
 			}
 
 			_clientArtifactId = 0;
-			testHelper = null;
+			_testHelper = null;
 			_servicesManager = null;
-			_client = null;
 		}
 
 		[Test]
 		public void CreateClientTest()
 		{
-			_clientArtifactId = Client.Create_Client(_client, _serviceFactory, _clientName);
+			_clientArtifactId = ClientHelper.CreateClient(_serviceFactory, _clientName);
 
 			Assert.Greater(_clientArtifactId, 0);
 		}
@@ -63,7 +54,7 @@ namespace Relativity.Test.Helpers.NUnit.Integration.ArtifactHelpers
 		[Test]
 		public void CreateClientTest_Failure()
 		{
-			Assert.Throws<AggregateException>(() => Client.Create_Client(_client, _serviceFactory, null));
+			Assert.Throws<Exception>(() => ClientHelper.CreateClient(_serviceFactory, null));
 		}
 
 		//helper method