+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 -------------------------
 
+## 2021-02-22
+- [REL-518419](https://jira.kcura.com/browse/REL-518419) - Removed RSAPI from Client class and renamed to ClientHelpers class. Also updated integration tests.
+
+-------------------------
+
 ## 2021-02-15
 - [REL-518412](https://jira.kcura.com/browse/REL-518412) - Removed NUnit dependency and deleted a constructor that accepted TestContext (an NUnit object).